+++ pkgs/test_core/lib/src/runner/configuration.dart
@@ -76,6 +76,10 @@ class Configuration {
   String get reporter => _reporter ?? defaultReporter;
   final String _reporter;
 
+  /// The map of file reporters where the key is the name of the reporter and
+  /// the value is the filepath to which its output should be written.
+  final Map<String, String> fileReporters;
+
   /// Whether to disable retries of tests.
   bool get noRetry => _noRetry ?? false;
   final bool _noRetry;
@@ -227,6 +231,7 @@ class Configuration {
       String configurationPath,
       String dart2jsPath,
       String reporter,
+      Map<String, String> fileReporters,
       String coverage,
       int pubServePort,
       int concurrency,
@@ -275,6 +280,7 @@ class Configuration {
         configurationPath: configurationPath,
         dart2jsPath: dart2jsPath,
         reporter: reporter,
+        fileReporters: fileReporters,
         coverage: coverage,
         pubServePort: pubServePort,
         concurrency: concurrency,
@@ -336,6 +342,7 @@ class Configuration {
       String configurationPath,
       String dart2jsPath,
       String reporter,
+      Map<String, String> fileReporters,
       String coverage,
       int pubServePort,
       int concurrency,
@@ -360,6 +367,7 @@ class Configuration {
         _configurationPath = configurationPath,
         _dart2jsPath = dart2jsPath,
         _reporter = reporter,
+        fileReporters = fileReporters ?? {},
         _coverage = coverage,
         pubServeUrl = pubServePort == null
             ? null
@@ -485,6 +493,7 @@ class Configuration {
         configurationPath: other._configurationPath ?? _configurationPath,
         dart2jsPath: other._dart2jsPath ?? _dart2jsPath,
         reporter: other._reporter ?? _reporter,
+        fileReporters: mergeMaps(fileReporters, other.fileReporters),
         coverage: other._coverage ?? _coverage,
         pubServePort: (other.pubServeUrl ?? pubServeUrl)?.port,
         concurrency: other._concurrency ?? _concurrency,