+++ pkgs/test/CHANGELOG.md
@@ -1,5 +1,7 @@
 ## 1.10.1-dev
 
+* Add `file_reporters` configuration option and `--file-reporter` CLI option to
+  allow specifying a separate reporter that writes to a file instead of stdout.
 * Internal cleanup.
 
 ## 1.10.0